made the My Account section its own component
